+++ lib/modules/theme/theme_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: theme
-// Updated at: 2026-04-17 17:31:45
-// Commit: fix(theme): prevent double tap duplicate action in theme
+// Updated at: 2026-04-17 19:10:33
+// Commit: test(theme): implement widget integration tests for theme
 
 class themeComponentState {
-  final String commitRef = '486';
-  final String lastMessage = 'fix(theme): prevent double tap duplicate action in theme';
-  final DateTime timestamp = DateTime.parse('2026-04-17T17:31:45+05:00');
+  final String commitRef = '491';
+  final String lastMessage = 'test(theme): implement widget integration tests for theme';
+  final DateTime timestamp = DateTime.parse('2026-04-17T19:10:33+05:00');
 }